+++ src/lib/version/changelog.ts
@@ -5,7 +5,7 @@ import { IContext, IVersionInfo } from "../../interface";
 
 export async function getChangelogVersion(context: IContext): Promise<IVersionInfo>
 {
-    let versionSystem: string;
+    let versionSystem: "auto" | "manual" | "semver" | "incremental";
     const version = await context.changelog.getVersion(context);
 
     if (!version)